Fixed validation test and fixed travis failing
