+++ .babelrc
@@ -1,3 +1,9 @@
 {
-  "presets": ["latest"]
+  "presets": [
+    ["env", {
+      "targets": {
+        "node": "current"
+      }
+    }]
+  ]
 }
