+++ package.json
@@ -5,6 +5,8 @@
   "main": "app.js",
   "dependencies": {
     "babel-cli": "^6.26.0",
+    "babel-polyfill": "^6.26.0",
+    "babel-preset-env": "^1.7.0",
     "babel-preset-latest": "^6.24.1",
     "bcrypt": "^3.0.0",
     "body-parser": "^1.18.3",
