+++ server/middlewares/mealChecker.js
@@ -2,7 +2,7 @@ import pool from '../database/connectdatabase';
 
 const mealNameCheck = (req, res, next) => {
   const { mealName } = req.body;
-  pool.query(`SELECT * FROM menu WHERE mealname = '${mealName}'`, (err, result) => {
+  pool.query(`SELECT * FROM menu WHERE mealname = '${mealName.trim()}'`, (err, result) => {
     if (result.rowCount > 0) {
       return res.status(400).json({
         status: 'failed',
